% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zeroes_to_HMDB.R
\name{add_zeroes_to_HMDB}
\alias{add_zeroes_to_HMDB}
\title{Ensure HMDB identifiers have 7 digits}
\usage{
add_zeroes_to_HMDB(string)
}
\arguments{
\item{string}{A single chr, beginning with "HMDB"}
}
\value{
A single chr, in the format "HMDBxxxxxxx"
}
\description{
PubChem search requires 7 digits in HMDB identifiers. This function repairs them.
}
\examples{
add_zeroes_to_HMDB("HMDB1")
# [1] "HMDB0000001"
}
