% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_complimentary_cols.R
\name{.coalesce_complimentary_cols}
\alias{.coalesce_complimentary_cols}
\title{Coalesce columns with complementary missingness}
\usage{
.coalesce_complimentary_cols(data, cols, output_col_name = "merged_col")
}
\arguments{
\item{data}{a dataframe with columns to be coalesced}

\item{cols}{a chr vector containing the names of columns to be coalesced. Must be the same type (e.g. \code{chr}).}

\item{output_col_name}{a name for the new column containing colasced data. Default is "merged_col"}
}
\value{
The coalesced column appended to \code{data}, minus the columns that were coalesced
}
\description{
Coalesces dataframe columns where there are NA values in unique locations
across columns. Designed to be used after left join, but will work on any
set of columns that have one entry per row and are of the same type.
}
\examples{
NA_3 <- rep(NA_character_, 3)

df_1 <- data.frame(
  a = c("A", "B", "C", NA_3, NA_3),
  b = c(NA_3, "D", "E", "F", NA_3),
  c = c(NA_3, NA_3, "G", "H", "I")
)

# Appends a new column containing coalesced data
.coalesce_complimentary_cols(df_1, cols = names(df_1))

# If there is mroe than one entry in any row, return NA for that row
df_2 <- df_1
df_2[1,3] <- "X"

.coalesce_complimentary_cols(df_2, cols = names(df_2))
}
