% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logic.R
\name{.get_logic}
\alias{.get_logic}
\title{Decide which IDs to query}
\usage{
.get_logic(data, col)
}
\arguments{
\item{data}{a dataframe}

\item{col}{chr. The name of a column in \code{data} which contains database
identifiers}
}
\value{
a logical vector
}
\description{
This is a helper function, which other functions use to work out which
database identifier to query.
}
