% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_and_bind.R
\name{.query_and_bind}
\alias{.query_and_bind}
\title{Send multiple queries to pubchem and return a dataframe}
\usage{
.query_and_bind(queries, query_fun, id_colname, label)
}
\arguments{
\item{queries}{chr vector. A set of identifiers (HMDB, LIPID-MAPS or CAS) or
metabolite names to query PubChem with}

\item{query_fun}{a PubChem querying function. Either \code{identifier_standardiser}
or \code{synonym_search}}

\item{id_colname}{chr. A name of a column where the queries are stored}

\item{label}{chr. The type of identifier ("CAS", "HMDB", "LMP" or "Name")}
}
\value{
a dataframe containing the PubChem CID, InChIKey, PubChem Title and
the identifier used for searching. Also includes PubChem error messages
}
\description{
This is designed to be a helper function which is repeatedly called via
\code{query_DB} to query the PubChem database.
}
