% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_PubChem_DB.R
\name{init_PubChem_cache}
\alias{init_PubChem_cache}
\title{Initialise a PubChem cache}
\usage{
init_PubChem_cache(path = NULL)
}
\arguments{
\item{path}{chr. A file path to store the database.
Default NULL will set the database location to "./Documents/PubChem_DB"}
}
\value{
An RDS environment
}
\description{
Initialise a PubChem cache
}
\examples{
cache <- init_PubChem_cache()    # Initialises cache in default location

filepath <- file.path(Sys.getenv("HOME"), "cache")
DB <- init_PubChem_cache(filepath)     # Initialises database in "./cache"
}
