% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_DB.R
\name{query_DB}
\alias{query_DB}
\title{Batch query the PubChem database for Title, CID and InChIKey}
\usage{
query_DB(data, cache)
}
\arguments{
\item{data}{dataframe. A metabolomics dataset from MassHunter explorer. Should
contain the columns:
\describe{
\item{Name}{chr. The putative metaobolite identification as output by MassHunter explorer}
\item{Formula}{chr. The chemcial formula of the metabolite}
\item{CAS}{chr. The CAS identifier}
\item{HMDB}{chr. The HMDB identifier}
\item{LMP}{chr. The LIPID-MAPS identifier}
}}

\item{cache}{an RDS cache environment. Generated with \code{init_PubChem_cache}.}
}
\value{
the original dataframe, appended with the InChIKey, PubChem title,
PubChem compound ID (CID), and any error messages arising from PubChem queries
}
\description{
Batch query the PubChem database for Title, CID and InChIKey
}
\section{}{

}

\examples{
cache <- init_PubChem_cache()

# Query the PubChem database
batch_query <- query_DB(
   data = explorer_metadata,
   cache = cache
)
}
