% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_pubchem.R
\name{query_pubchem}
\alias{query_pubchem}
\title{Batch query the PubChem database for Title, CID and InChIKey}
\usage{
query_pubchem(
  data,
  cache,
  name = "Name",
  cas = "CAS",
  lipidmaps = "LMP",
  hmdb = "HMDB"
)
}
\arguments{
\item{data}{dataframe. A metabolomics dataset from MassHunter explorer. To
use default settings, your dataframe should contain the columns:
\describe{
\item{Name}{chr. The putative metaobolite identification as output by MassHunter explorer}
\item{CAS}{chr. The CAS identifier}
\item{HMDB}{chr. The HMDB identifier}
\item{LMP}{chr. The LIPID-MAPS identifier}
}}

\item{cache}{an RDS cache environment. Generated with \code{init_PubChem_cache}.}

\item{name}{chr. The name of a column in \code{data} to override looking for
metabolite names in the default column "Name". Set to \code{NULL} to skip querying
PubChem with names.}

\item{cas}{chr. Change the "CAS" column. Behaves the same as \code{name}.}

\item{lipidmaps}{chr. Change the "LMP" column. Behaves the same as \code{name}.}

\item{hmdb}{chr. Change the "HMDB" column. Behaves the same as \code{name}.}
}
\value{
the original dataframe, appended with the InChIKey, PubChem title,
PubChem compound ID (CID), and any error messages arising from PubChem queries
}
\description{
\code{query_pubchem} is designed to work with MassHunter Explorer outputs, but can
receive any dataframe containing columns with metabolite names, CAS identifiers,
HMDB identifiers, or LIPID-MAPS identifiers. The function systematically
queries the PubChem database using the PUG-REST API and returns harmonised
names (the PubChem Title), PubChem compound IDs (CIDs) and an InChIKey.
}
\examples{
# Set up a cache in the default location
cache <- init_PubChem_cache()

# Query the PubChem database
batch_query <- query_pubchem(
   data = explorer_metadata,
   cache = cache
)
# Query the PubChem database with metabolite names only
name_query <- query_pubchem(
   data = explorer_metadata,
   cache = cache,
   cas = NULL, lipidmaps = NULL, hmdb = NULL
)
}
