% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonym_search.R
\name{synonym_search}
\alias{synonym_search}
\title{Standardise metabolite names}
\usage{
synonym_search(name, cache)
}
\arguments{
\item{name}{chr. A metabolite name to search for.}

\item{cache}{an RDS cache environment. Generated with \code{init_PubChem_cache}.}
}
\value{
a one-row dataframe with the PubChem compound ID, InChIKey, and PubChem title
}
\description{
Standardises a metabolite name to the PubChem CID, InChIKey
and title using PubChem synonym search.
Returns a single row dataframe with these as outputs.
Requires internet to query the PubChem database.
}
\examples{
# Set up a cache in the default location
cache <- init_PubChem_cache()

# Search using PubChem synonyms
syn_id <- synonym_search(
  name = "Alanine",
  cache = cache)
}
