% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_me.R
\name{classify_me}
\alias{classify_me}
\title{Batch query the ClassyFire metabolite classification database}
\usage{
classify_me(inchikeys, cache = NULL)
}
\arguments{
\item{inchikeys}{chr vector containing only InChI keys and \code{NA}}

\item{cache}{the file path of an SQLite cache. Default \code{NULL} places the
cache in the user's home folder.}
}
\value{
a dataframe with columns containing the InChI key and the first four
classification levels from the ClassyFire database
}
\description{
Receives a vector in InChI keys and queries the ClassyFire metabolite
classification database. Returns a simple dataframe with results, ready for
merging or further analyses. This function wraps functions from the
\code{classyfireR} package.
}
\examples{
# Some example InChI keys to test
keys <- c(
  "BACDZNLMIXNCOG-JTCWOHKRSA-N",
  "TVHAWOPAFXXIQM-PLNGDYQASA-N",
  "ROFVXGGUISEHAM-UHFFFAOYSA-N",
  "KQTROLJZWNACJT-VKAVYKQESA-N",
  "GMYNCKRSFMEXPG-UHFFFAOYSA-N")

# Initialise a cache in the default location
classes <- classify_me(keys)

# Future calls are faster as they access the local cache instead of the
# online classyfire cache
classes <- classify_me(keys)

# You can setup your cache wherever you like
classes_alt <- classify_me(
  keys,
 cache = file.path(Sys.getenv("HOME"), "Metabolomics", "classyfire_DB/ClassyFireCache.db")
)
}
