% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_standardiser.R
\name{identifier_standardiser}
\alias{identifier_standardiser}
\title{Standardise CAS, LIPID-MAPS and HMDB IDs}
\usage{
identifier_standardiser(identifier, type = c("CAS", "LMP", "HMDB"), cache)
}
\arguments{
\item{identifier}{a database identifier to search for (e.g. "58-18-4", "LMFA01100004" or "HMDB0002100")}

\item{type}{the database you are searching. Either "CAS", "LMP", or "HMDB"}

\item{cache}{an RDS cache environment. Generated with \code{init_PubChem_cache}.}
}
\value{
a one-row dataframe with the PubChem compound ID, InChIKey, and PubChem title
}
\description{
Standardises identifiers (CAS, HMDB, LIPID-MAPS) to the PubChem CID, InChIKey
and title. Returns a single row dataframe with these as outputs.
Requires internet to query the PubChem database.

HMDB identifiers must be in the format "HMDBxxxxxxx" with seven digits.
A helper function, \code{add_zeroes_to_HMDB} is available to coerce HMDB IDs to this format.
}
\examples{
# Set up a cache in the default location
cache <- init_PubChem_cache()

# Search for a CAS identifier
cas_id <- identifier_standardiser(
  identifier = "58-18-4",
  type = "CAS",
  cache = cache)

# Search for a LIPID-MAPS identifier
lipidmap_id <- identifier_standardiser(
  identifier = "LMFA01100004",
  type = "LMP",
  cache = cache)

# Search for an HMDB identifier
hmdb_id <- identifier_standardiser(
  identifier = "HMDB0002100",
  type = "HMDB",
  cache = cache)
}
